Added additional comments and changed software serial ports
Expanded the header and program flow comments to be more clear and
included hardware pin connections.

Changed the software serial port from pin 49 to pin 13 so that the same
firmware will work on the Uno and the Mega without alterations. The
platform ships with either microcontroller depending on the options
selected during checkout.